All codes/uv/pyproject.toml
[project]
name = "uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=2.0.1",
    "google-generativeai>=0.8.6",
    "groq>=1.2.0",
    "openai-agents>=0.17.0",
    "python-dotenv>=1.2.2",
]
